refactor: move helpers into library file helper
